Make minor alterations to scene
Added a staircase to the scene to allow AI pathing to the hgh areas later on.  Made sure all areas of the default map are accessible for now.  The layout will most likely change later on, but it's important at this prototyping stage to have a fully accessible play space.