Add AIPerception component to the NPC_AI_BP
Added an AI Sight config to the AIPerception component to allow the guards to perceive the player in the layer.  Set reasonable parameters for the sight and lose sight radii, and also checked all detection by affiliation parameters, as at this point I'm unsure as to the relationship between the player character and the guards.