Add patrol component and NPC_AI_BP to the character blueprint from the Animation Starter Pack
Allows the blueprint in the starter pack to be used as the new patrolling guard asset (and is renamed to PatrollingGuard2).  The reason for this is because there are a lot of events already configured in this blueprint (such as walking, crouching, etc), that can be utilised in this game.  Rather than having to set them up again in the original PatrollingGuard asset, it makes sense to just repurpose the character blueprint imported with the animation starter pack.