Make the guard investigate the player's last known location
Create a service that updates the player last known location as a blackboard key.  This is continuously set while the Enemy blackboard key is set, then when the guard loses sight of the player, the guard investigates this last known location and pauses for a few seconds before resuming his patrol route.