Bind the "Fire" action to the OnFire() function
This has to be done in BeginPlay() as the SetupPlayerInputComponent() overrides the Pawn's function, which means that these bindings are setup at construction.  As the gun is not constucted at this point, the binding of the Gun's OnFire() method needs to be done later, which is why it is called in BeginPlay().