Prevent projectile collisions with Locking/Destroy Volumes
The default collision settings for the Locking and Destroy Volumes generate hits on collision with projectiles, rather than letting them pass through.  This results in the projectiles being destroyed when coming into contact with the Volumes.  Changing the collision presets to "Trigger" on both volumes now correctly results in projectiles being able to pass though.