Fix bug which made the guards ignore min. approach distance
This bug was introduced when the "Add noise" componenet was added to the Weapon.  Unfortunately, it transpired that the NPC_AI_BP was too complex to debug this issue, so the "Add noise" component has been removed so that the NPC_AI_BP can be refactored first.  After the refactor, it should be easier to troubleshoot the approach distance issue when adding in the "Add noise" component.