Fix issue with guards not following the player
If a guard sees a player, then he shoots at the player while moving within the designated target radius, but if the player moves further away, then the guard did not follow the player while shooting.  This has been fixed, so the guards will now chase the player while continuously shooting at them until the player is no longer in line of sight.