Fixed issue with guards not shooting
For some reason, the calculated distance between the guards and the enemies is greater than the 300cm acceptance radius, this results in the guard not shooting, or moving towards the player.  I think this is because the accepted distance is calculated between the two closest points, whereas the way it is caluclated within IsWithiRange is between the central points of the guard and player.  Upping the radius in IsWithinRange from 300cm to 400cm seems to fix the issue, as the guard stops moving towards the player when the distance is approsimately 397cm.