Fix error message when guard chases player
When the guard loses sight of the player, it appears that the blackboard key is cleared which results in an error.  I'm not sure why this is, as this logic should only execute when the Enemy key is definitely set.  Adding in an IsValid node prevents the error, and is a similar fix that was employed in BTTask_UpdateLastSeenLocation.