Fix issue with spawning in filled spaces
Actors of the same type were being spawned simultaneously via a TArray.  When these spawn locations are added to the TArray, they only check if the space is filled by another Actor already in the world.  This means that objects of the same type can spawn over one another.  This is alleviated by spawning each object into the world, and then checking for Candidate positions, before spawning the next object.  The code has also been refactored to take advantage of templates, so that the AI and props can both use the same function call to RandomlyPlaceActors().